backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
msal>=1.28
httpx>=0.27
openai>=1.30
gunicorn>=22.0
uvicorn[standard]>=0.30
psycopg[binary]>=3.1
dj-database-url>=2.1
python-decouple>=3.8
whitenoise>=6.7
